refactored floatpicker input

+++ FilterPlayground/Attributes View/KernelArgumentValueView.swift
@@ -10,7 +10,7 @@ import UIKit
 
 protocol KernelArgumentValueView: class {
     // 0 for default height
-    var prefferedHeight: Float { get }
+    var prefferedHeight: CGFloat { get }
     var prefferedUIAxis: UILayoutConstraintAxis { get }
     var updatedValueCallback: ((KernelArgumentValue) -> Void)? { get set }
     var value: KernelArgumentValue { get set }
